requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
altair==4.1.0
arrow==0.15.6
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
binaryornot==0.4.4
bleach==3.1.5
blinker==1.4
boto3==1.14.23
botocore==1.17.23
cachetools==4.1.1
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1598370813909/work
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233707986/work
olefile==0.46
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1594213010297/work
poyo==0.5.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyarrow==0.17.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pydeck==0.4.0b2
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2020.1
pyzmq==19.0.1
requests==2.23.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.5.1
Send2Trash==1.5.0
six==1.15.0
streamlit==0.63.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
torch==1.3.1+cpu
torchvision==0.4.2+cpu
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.9
validators==0.16.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
